+++ flutter/test/default_integrations_test.dart
@@ -92,7 +92,7 @@ void main() {
 
     final integrationA = FlutterErrorIntegration();
     integrationA.call(fixture.hub, fixture.options);
-    integrationA.close();
+    await integrationA.close();
 
     final integrationB = FlutterErrorIntegration();
     integrationB.call(fixture.hub, fixture.options);
@@ -104,7 +104,7 @@ void main() {
     expect(numberOfDefaultCalls, 1);
   });
 
-  test('FlutterErrorIntegration close restored default onError', () {
+  test('FlutterErrorIntegration close restored default onError', () async {
     final defaultOnError = (FlutterErrorDetails errorDetails) async {};
     FlutterError.onError = defaultOnError;
 
@@ -112,12 +112,12 @@ void main() {
     integration.call(fixture.hub, fixture.options);
     expect(false, defaultOnError == FlutterError.onError);
 
-    integration.close();
+    await integration.close();
     expect(FlutterError.onError, defaultOnError);
   });
 
   test('FlutterErrorIntegration default not restored if set after integration',
-      () {
+      () async {
     final defaultOnError = (FlutterErrorDetails errorDetails) async {};
     FlutterError.onError = defaultOnError;
 
@@ -128,7 +128,7 @@ void main() {
     final afterIntegrationOnError = (FlutterErrorDetails errorDetails) async {};
     FlutterError.onError = afterIntegrationOnError;
 
-    integration.close();
+    await integration.close();
     expect(FlutterError.onError, afterIntegrationOnError);
   });